TST: use np.int64 in sparse groupby tests for 32-bit compat
On 32-bit platforms (Linux-32-bit, Pyodide/wasm32), SparseDtype(int, ...)
resolves to int32 while DataFrame int columns are int64, causing dtype
mismatches. Use np.int64 explicitly.

+++ pandas/tests/groupby/test_sparse.py
@@ -26,8 +26,8 @@ def sparse_df(fill_value):
         }
     )
     sparse = dense.copy()
-    sparse["val1"] = sparse["val1"].astype(SparseDtype(int, fill_value))
-    sparse["val2"] = sparse["val2"].astype(SparseDtype(int, fill_value))
+    sparse["val1"] = sparse["val1"].astype(SparseDtype(np.int64, fill_value))
+    sparse["val2"] = sparse["val2"].astype(SparseDtype(np.int64, fill_value))
     return dense, sparse
 
 
@@ -103,7 +103,7 @@ class TestSparseGroupby:
         vals = [1, 0, 0, 3, 0, 5]
         keys = ["a", "a", "b", "b", "a", "b"]
         dense_ser = Series(vals, name="val")
-        sparse_ser = dense_ser.astype(SparseDtype(int, fill_value))
+        sparse_ser = dense_ser.astype(SparseDtype(np.int64, fill_value))
         result = sparse_ser.groupby(keys).mean()
         expected = dense_ser.groupby(keys).mean()
         tm.assert_series_equal(result, expected)
